requirements.txt
beautifulsoup4>=4.12
markdownify>=0.11
python-slugify>=8.0
python-dotenv>=1.0
requests>=2.31
